+++ lib/profile.js
@@ -182,11 +182,11 @@ function mark(source) {
 }
 
 function isShortDateStr(str) { //example:2008-07-22
-    var dateFormat = /^\d{4}-\d{2}-\d{2}$/;
-    if (dateFormat.test(str)) {
-        return true;
-    } else {
-        return false;
-    }
+  var dateFormat = /^\d{4}-\d{2}-\d{2}$/;
+  if (dateFormat.test(str)) {
+    return true;
+  } 
+  return false;
+    
 }
 module.exports = { getProfile, getProfileFromFile, mark, isShortDateStr };
